[Fix] record-/list 빌드 에러 수정
